backend/requirements.txt
fastapi
uvicorn
boto3
pymupdf
python-docx
faiss-cpu
numpy
Pillow
python-multipart
python-dotenv
